+++ src/Redux/alerts.js
@@ -45,16 +45,15 @@ export default function reducer(state=defaultState, action){
 /* ======= Action creators =========
 *
 */
+const _clearTimeInterval = () => ({ type: CLEAR_TIME_INTERVAL }) // helper action creator that shouldn't be used outside
 
 export const showMsg = () => ({ type: SHOW_MSG})
-export const clearMsg = () => ({ type: CLEAR_MSG})
-export const clearTimeInterval = () => ({ type: CLEAR_TIME_INTERVAL })
 
-// Helper Function
-// const delay = (ms) => new Promise(resolve => setTimeout(resolve, ms))
+export const clearMsg = () => ({ type: CLEAR_MSG})
+
 export const delayedClearMsg = (ms) => (dispatch, getState) => {
   if (getState().alerts.timeInterval){
-    dispatch(clearTimeInterval())
+    dispatch(_clearTimeInterval())
   }
   const timeInterval = setTimeout(()=>{ dispatch(clearMsg())}, ms)
   dispatch({ type: SET_TIME_INTERVAL, payload: timeInterval })