+++ CHANGELOG.md
@@ -50,3 +50,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 - combined store.js and reducer.js into a single `index.js` file in the `Redux/` directory
 - changed file paths in all the 'smart' components directing to the store
+
+### Added 
+- added alerts reducer (shows msg, clears msg, updates msg, sets time interval for clearing)